Remove grpc-js depdency from protbufs package
