+++ packages/rpc/src/schemas/messages/hub_event.proto
