+++ packages/rpc/src/schemas/messages/hub_state.proto
