+++ packages/rpc/src/schemas/messages/id_registry_event.proto
