+++ packages/rpc/src/schemas/messages/name_registry_event.proto
